Finished ReplicationManager and got working exemple

+++ utils/frameworkTP3.cpp
@@ -1,6 +1,6 @@
 #include "frameworkTP3.h"
 
-float uqac::utils::randomInt(int min, int max)
+int uqac::utils::randomInt(int min, int max)
 {
     return rand() % max + min;
 }
